[ALPS04713825] iommu: fix code defect
[Detail]
free before error return
correct argument type

MTK-Commit-Id: 6a52661bcf12b4d189d1cf01e5f9ae6e538dc77f

CR-Id: ALPS04713825
Feature: [Module]ION/M4U

+++ drivers/staging/android/ion/ion_system_heap.c
@@ -92,8 +92,8 @@ static void free_buffer_page(struct ion_system_heap *heap,
 	if (buffer->private_flags & ION_PRIV_FLAG_SHRINKER_FREE) {
 		__free_pages(page, order);
 		if (atomic64_sub_return((1 << order), &page_sz_cnt) < 0) {
-			seq_printf(NULL, "underflow!, total[%lu]free[%lu]\n",
-				   (unsigned long)atomic64_read(&page_sz_cnt),
+			seq_printf(NULL, "underflow!, total[%llu]free[%lu]\n",
+				   atomic64_read(&page_sz_cnt),
 				   (unsigned long)(1 << order));
 			atomic64_set(&page_sz_cnt, 0);
 		}
