+++ drivers/staging/android/ion/mtk/ion_mm_heap.c
@@ -170,8 +170,8 @@ static void free_buffer_page(struct ion_system_heap *heap,
 	} else {
 		__free_pages(page, order);
 		if (atomic64_sub_return((1 << order), &page_sz_cnt) < 0) {
-			IONMSG("underflow!, total_now[%lu]free[%lu]\n",
-			       (unsigned long)atomic64_read(&page_sz_cnt),
+			IONMSG("underflow!, total_now[%llu]free[%lu]\n",
+			       atomic64_read(&page_sz_cnt),
 			       (unsigned long)(1 << order));
 			atomic64_set(&page_sz_cnt, 0);
 		}
@@ -1167,7 +1167,7 @@ static int ion_mm_heap_debug_show(struct ion_heap *heap, struct seq_file *s,
 	}
 	current_ts = sched_clock();
 	do_div(current_ts, 1000000);
-	IONMSG("current time %lld ms, total: %lu!!\n",
+	IONMSG("current time %lld ms, total: %llu!!\n",
 	       current_ts, atomic64_read(&page_sz_cnt) * 4096);
 #ifdef CONFIG_MTK_IOMMU_V2
 	mtk_iommu_log_dump(s);
